refactor: split main.ts into logical modules and add documentation
- Created `electron/src/` with `types.ts`, `config.ts`, `font-service.ts`, `ssh-manager.ts`, `update-service.ts`, and `ipc-handlers.ts`.
- Refactored `electron/main.ts` to be a clean entry point.
- Added JSDoc comments in Russian to all new functions and modules.
- Created `docs/ARCHITECTURE.md` (in Russian) explaining the new structure.
- Improved type safety in SSH and Update logic (removed `any`).
- Verified build and linting.
