+++ docs/ARCHITECTURE.md
@@ -1,0 +1,45 @@
+# Архитектура основного процесса (Main Process)
+
+Данный документ описывает структуру и организацию кода в директории `electron/`, отвечающей за работу основного процесса приложения YetAnotherSSHClient.
+
+## Структура файлов
+
+После рефакторинга код разделен на логические модули для обеспечения удобства поддержки и масштабируемости.
+
+```text
+electron/
+├── main.ts             # Точка входа, жизненный цикл приложения, создание окна
+├── preload.ts          # Скрипт моста между Electron и Renderer процессом
+└── src/                # Модули логики основного процесса
+    ├── types.ts        # Общие интерфейсы и типы данных
+    ├── config.ts       # Работа с конфигурационным файлом (.minissh_config.json)
+    ├── font-service.ts # Получение списка системных шрифтов
+    ├── ssh-manager.ts  # Управление SSH-сессиями и потоками
+    ├── update-service.ts # Проверка обновлений через GitHub API
+    └── ipc-handlers.ts # Регистрация и обработка IPC-событий
+```
+
+## Описание модулей
+
+### 1. Main Entry (`main.ts`)
+Отвечает за инициализацию Electron, создание главного окна браузера (`BrowserWindow`) и управление глобальными событиями приложения (запуск второго экземпляра, выход).
+
+### 2. Config (`src/config.ts`)
+Обеспечивает сохранение и загрузку пользовательских настроек (размер окна, избранные серверы, шрифты). Конфигурация хранится в домашней директории пользователя в формате JSON.
+
+### 3. SSH Manager (`src/ssh-manager.ts`)
+Инкапсулирует состояние активных SSH-соединений. Хранит карты (Maps) для объектов `ssh2.Client`, `ssh2.ClientChannel` (shell) и сетевых сокетов. Предоставляет методы для очистки ресурсов при закрытии вкладок или выходе из приложения.
+
+### 4. Font Service (`src/font-service.ts`)
+Выполняет системные команды для получения списка установленных шрифтов в зависимости от операционной системы (Windows/macOS/Linux).
+
+### 5. Update Service (`src/update-service.ts`)
+Раз в сутки проверяет наличие новых релизов в репозитории GitHub. Сравнивает версии и отправляет уведомление во фронтенд.
+
+### 6. IPC Handlers (`src/ipc-handlers.ts`)
+Центральное место регистрации всех обработчиков межпроцессного взаимодействия. Делегирует выполнение бизнес-логики соответствующим модулям и возвращает результаты во фронтенд через `ipcMain.handle` или `event.reply`.
+
+## Принципы разработки
+- **Модульность**: Каждая функция должна находиться в соответствующем сервисе.
+- **Документация**: Весь экспортируемый функционал должен быть покрыт JSDoc комментариями.
+- **Безопасность**: Доступ к Node.js API из рендерера разрешен только через ограниченный `preload.ts` и типизированные IPC-вызовы.
