+++ electron/main.ts
@@ -1,129 +1,23 @@
-import { app, BrowserWindow, ipcMain, shell, IpcMainEvent, dialog } from 'electron'
+import { app, BrowserWindow } from 'electron'
 import * as path from 'node:path'
-import { Client, PseudoTtyOptions, type ClientChannel, type ConnectConfig } from 'ssh2'
-import * as fs from 'node:fs'
 import { fileURLToPath } from 'node:url'
-import * as os from 'node:os'
-import * as net from 'node:net'
-import { exec } from 'node:child_process'
-import { promisify } from 'node:util'
-
-const execAsync = (cmd: string) => promisify(exec)(cmd, { maxBuffer: 1024 * 1024 * 10 })
-
-/* ================= TYPES ================= */
-
-interface SSHConfig {
-    id?: string
-    name: string
-    user: string
-    host: string
-    port: number
-    password?: string
-    authType?: 'password' | 'key'
-    privateKeyPath?: string
-    osPrettyName?: string
-}
-
-interface AppConfig {
-    terminalFontName: string
-    terminalFontSize: number
-    uiFontName: string
-    uiFontSize: number
-    theme: string
-    favorites: SSHConfig[]
-    x: number
-    y: number
-    width: number
-    height: number
-    maximized: boolean
-    lastUpdateCheck?: number
-}
-
-interface SshConnectPayload {
-    id: string
-    config: SSHConfig
-    cols?: number
-    rows?: number
-}
+import { loadConfig, saveConfig } from './src/config.js'
+import { cleanupAll } from './src/ssh-manager.js'
+import { checkUpdates } from './src/update-service.js'
+import { registerIpcHandlers } from './src/ipc-handlers.js'
 
 /* ================= INIT ================= */
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url))
-const configPath = path.join(os.homedir(), '.minissh_config.json')
-
 let mainWindow: BrowserWindow | null = null
 
-const sshClients = new Map<string, Client>()
-const shellStreams = new Map<string, ClientChannel>()
-const sshSockets = new Map<string, net.Socket>()
-
-/* ================= CONFIG ================= */
-
-const DEFAULT_CONFIG: AppConfig = {
-    terminalFontName: 'JetBrains Mono',
-    terminalFontSize: 17,
-    uiFontName: 'JetBrains Mono',
-    uiFontSize: 12,
-    theme: 'Gruvbox Light',
-    favorites: [],
-    x: 353,
-    y: 141,
-    width: 1254,
-    height: 909,
-    maximized: false,
-    lastUpdateCheck: 0
-}
-
-function loadConfig(): AppConfig {
-    if (!fs.existsSync(configPath)) return DEFAULT_CONFIG
-    try {
-        return JSON.parse(fs.readFileSync(configPath, 'utf-8')) as AppConfig
-    } catch {
-        return DEFAULT_CONFIG
-    }
-}
-
-function saveConfig(config: AppConfig): void {
-    fs.writeFileSync(configPath, JSON.stringify(config, null, 2))
-}
-
-async function getSystemFonts(): Promise<string[]> {
-    const fallbacks = [
-        'JetBrains Mono', 'Consolas', 'Courier New', 'Segoe UI',
-        'Roboto', 'Ubuntu Mono', 'Arial', 'monospace', 'sans-serif'
-    ]
-    try {
-        let fonts: string[] = []
-        if (process.platform === 'win32') {
-            const cmd = `powershell -NoProfile -Command "[System.Reflection.Assembly]::LoadWithPartialName('System.Drawing') | Out-Null; (New-Object System.Drawing.Text.InstalledFontCollection).Families.Name"`
-            const { stdout } = await execAsync(cmd)
-            fonts = stdout.split(/\r?\n/)
-                .map(s => s.trim())
-                .filter(Boolean)
-        } else if (process.platform === 'darwin') {
-            try {
-                const { stdout } = await execAsync('atsutil font -list | grep "^\\s*Family:" | awk -F ": " \'{print $2}\'')
-                fonts = stdout.split(/\r?\n/).map(s => s.trim()).filter(Boolean)
-            } catch {
-                const { stdout } = await execAsync('system_profiler SPFontsDataType | grep "Family:" | awk -F ": " \'{print $2}\'')
-                fonts = stdout.split(/\r?\n/).map(s => s.trim()).filter(Boolean)
-            }
-        } else {
-            const { stdout } = await execAsync('fc-list : family')
-            fonts = stdout.split(/\r?\n/)
-                .flatMap(s => s.split(','))
-                .map(s => s.trim())
-                .filter(Boolean)
-        }
-        return Array.from(new Set([...fallbacks, ...fonts])).sort()
-    } catch (e) {
-        console.error('Failed to get system fonts:', e)
-        return fallbacks.sort()
-    }
-}
-
-/* ================= WINDOW ================= */
-
+/**
+ * Возвращает цвет фона окна в зависимости от выбранной темы.
+ * Используется для предотвращения белой вспышки при загрузке.
+ *
+ * @param {string} theme - Название темы.
+ * @returns {string} Hex-код цвета фона.
+ */
 function getThemeColor(theme: string): string {
     switch (theme) {
         case 'Dark': return '#1e1e1e'
@@ -132,6 +26,9 @@ function getThemeColor(theme: string): string {
     }
 }
 
+/**
+ * Создает основное окно приложения.
+ */
 function createWindow(): void {
     const config = loadConfig()
 
@@ -156,6 +53,10 @@ function createWindow(): void {
 
     let saveTimeout: NodeJS.Timeout | null = null
 
+    /**
+     * Сохраняет состояние окна (размеры, положение) в конфигурацию.
+     * Использует debounce (500мс) для оптимизации.
+     */
     const saveWindowState = () => {
         if (saveTimeout) clearTimeout(saveTimeout)
         saveTimeout = setTimeout(() => {
@@ -190,6 +91,7 @@ function createWindow(): void {
 
 /* ================= APP LIFECYCLE ================= */
 
+// Обработка запуска одного экземпляра приложения
 if (!app.requestSingleInstanceLock()) {
     app.quit()
 } else {
@@ -204,8 +106,14 @@ if (!app.requestSingleInstanceLock()) {
         if (process.platform === 'win32') {
             app.setAppUserModelId('com.yash.client')
         }
+
+        // Регистрация обработчиков IPC
+        registerIpcHandlers(() => mainWindow)
+
         createWindow()
-        setTimeout(checkUpdates, 5000)
+
+        // Отложенная проверка обновлений
+        setTimeout(() => checkUpdates(mainWindow), 5000)
     })
 
     app.on('before-quit', cleanupAll)
@@ -213,193 +121,4 @@ if (!app.requestSingleInstanceLock()) {
     app.on('window-all-closed', () => {
         if (process.platform !== 'darwin') app.quit()
     })
-}
-
-function cleanupAll(): void {
-    shellStreams.forEach(s => s.destroy())
-    sshClients.forEach(c => c.destroy())
-    sshSockets.forEach(s => s.destroy())
-    shellStreams.clear()
-    sshClients.clear()
-    sshSockets.clear()
-}
-
-/* ================= IPC ================= */
-
-ipcMain.handle('get-config', () => loadConfig())
-ipcMain.handle('get-system-fonts', () => getSystemFonts())
-ipcMain.handle('save-config', (_, config: AppConfig) => saveConfig(config))
-
-ipcMain.handle('select-key-file', async () => {
-    const { canceled, filePaths } = await dialog.showOpenDialog({
-        properties: ['openFile'],
-        filters: [
-            { name: 'Keys', extensions: ['*', 'pem', 'ppk'] },
-            { name: 'All Files', extensions: ['*'] }
-        ]
-    })
-    if (canceled) return null
-    return filePaths[0]
-})
-
-ipcMain.on('ssh-connect', (event: IpcMainEvent, payload: SshConnectPayload) => {
-    const { id, config, cols = 80, rows = 24 } = payload
-
-    sshSockets.get(id)?.destroy()
-    sshClients.get(id)?.destroy()
-    shellStreams.delete(id)
-    sshClients.delete(id)
-    sshSockets.delete(id)
-
-    const sshClient = new Client()
-    sshClients.set(id, sshClient)
-
-    const socket = net.connect(config.port || 22, config.host)
-    sshSockets.set(id, socket)
-
-    socket.on('connect', () => {
-        socket.setNoDelay(true)
-
-        const connectConfig: ConnectConfig = {
-            sock: socket,
-            username: config.user,
-            readyTimeout: 20000
-        }
-
-        if (config.authType === 'key' && config.privateKeyPath) {
-            try {
-                connectConfig.privateKey = fs.readFileSync(config.privateKeyPath)
-            } catch (err: any) {
-                event.reply(`ssh-error-${id}`, `Failed to read private key: ${err.message}`)
-                cleanupConnection(id)
-                return
-            }
-        } else {
-            connectConfig.password = Buffer.from(config.password ?? '', 'base64').toString('utf8')
-        }
-
-        sshClient.connect(connectConfig)
-    })
-
-    socket.on('error', (err: Error) => {
-        event.reply(`ssh-error-${id}`, `Socket error: ${err.message}`)
-        cleanupConnection(id)
-    })
-
-    sshClient.on('ready', () => {
-        event.reply(`ssh-status-${id}`, 'Установлено SSH-соединение')
-
-        const pty: PseudoTtyOptions = { rows, cols, term: 'xterm-256color' }
-
-        sshClient.shell(pty, (err, stream) => {
-            if (err || !stream) {
-                event.reply(`ssh-error-${id}`, err?.message ?? 'Shell error')
-                return
-            }
-
-            shellStreams.set(id, stream)
-
-            stream.on('data', (chunk: Buffer) => {
-                event.reply(`ssh-output-${id}`, chunk)
-            })
-
-            stream.on('close', () => {
-                sshClient.end()
-                event.reply(`ssh-status-${id}`, 'SSH-соединение закрыто')
-            })
-        })
-    })
-
-    sshClient.on('error', (err: Error) => {
-        event.reply(`ssh-error-${id}`, err.message)
-        cleanupConnection(id)
-    })
-})
-
-ipcMain.on('ssh-input', (_, payload: { id: string; data: string }) => {
-    shellStreams.get(payload.id)?.write(payload.data)
-})
-
-ipcMain.on('ssh-resize', (_, payload: { id: string; cols: number; rows: number }) => {
-    shellStreams.get(payload.id)?.setWindow(payload.rows, payload.cols, 0, 0)
-})
-
-ipcMain.on('ssh-get-os-info', (event: IpcMainEvent, id: string) => {
-    const client = sshClients.get(id)
-    if (client) {
-        client.exec('cat /etc/os-release', (err, stream) => {
-            if (err) return
-            let output = ''
-            stream.on('data', (data: Buffer) => {
-                output += data.toString()
-            }).on('close', () => {
-                event.reply(`ssh-os-info-${id}`, output)
-            })
-        })
-    }
-})
-
-ipcMain.on('ssh-close', (_, id: string) => cleanupConnection(id))
-
-function cleanupConnection(id: string): void {
-    shellStreams.get(id)?.destroy()
-    sshClients.get(id)?.destroy()
-    sshSockets.get(id)?.destroy()
-    shellStreams.delete(id)
-    sshClients.delete(id)
-    sshSockets.delete(id)
-}
-
-ipcMain.on('window-minimize', () => mainWindow?.minimize())
-ipcMain.on('window-maximize', () => mainWindow?.isMaximized() ? mainWindow.unmaximize() : mainWindow?.maximize())
-ipcMain.on('window-close', () => {
-    cleanupAll()
-    mainWindow?.destroy()
-    app.exit(0)
-})
-
-ipcMain.on('open-external', (_, url: string) => shell.openExternal(url))
-
-async function checkUpdates() {
-    const config = loadConfig()
-    const now = Date.now()
-    const ONE_DAY = 24 * 60 * 60 * 1000
-
-    if (config.lastUpdateCheck && (now - config.lastUpdateCheck < ONE_DAY)) {
-        return
-    }
-
-    try {
-        const GITHUB_API_URL = "https://api.github.com/repos/megoRU/YetAnotherSSHClient/releases/latest"
-        const response = await fetch(GITHUB_API_URL, {
-            headers: { 'User-Agent': 'YetAnotherSSHClient' }
-        })
-        if (!response.ok) return
-
-        const data: any = await response.json()
-        const latestVersion = data.tag_name.replace(/^v/, '')
-        const currentVersion = app.getVersion()
-
-        if (isNewerVersion(latestVersion, currentVersion)) {
-            mainWindow?.webContents.send('update-available', {
-                version: latestVersion,
-                url: data.html_url
-            })
-        }
-
-        config.lastUpdateCheck = now
-        saveConfig(config)
-    } catch (err) {
-        console.error('Failed to check for updates:', err)
-    }
-}
-
-function isNewerVersion(latest: string, current: string): boolean {
-    const l = latest.split('.').map(Number)
-    const c = current.split('.').map(Number)
-    for (let i = 0; i < 3; i++) {
-        if (l[i] > (c[i] || 0)) return true
-        if (l[i] < (c[i] || 0)) return false
-    }
-    return false
 }
