+++ electron/src/config.ts
@@ -1,0 +1,47 @@
+import * as fs from 'node:fs'
+import * as path from 'node:path'
+import * as os from 'node:os'
+import { AppConfig } from './types.js'
+
+/** Путь к файлу конфигурации в домашней директории пользователя */
+export const configPath = path.join(os.homedir(), '.minissh_config.json')
+
+/** Конфигурация по умолчанию */
+export const DEFAULT_CONFIG: AppConfig = {
+    terminalFontName: 'JetBrains Mono',
+    terminalFontSize: 17,
+    uiFontName: 'JetBrains Mono',
+    uiFontSize: 12,
+    theme: 'Gruvbox Light',
+    favorites: [],
+    x: 353,
+    y: 141,
+    width: 1254,
+    height: 909,
+    maximized: false,
+    lastUpdateCheck: 0
+}
+
+/**
+ * Загружает конфигурацию из файла.
+ * Если файл не существует или поврежден, возвращает конфигурацию по умолчанию.
+ *
+ * @returns {AppConfig} Объект конфигурации приложения.
+ */
+export function loadConfig(): AppConfig {
+    if (!fs.existsSync(configPath)) return DEFAULT_CONFIG
+    try {
+        return JSON.parse(fs.readFileSync(configPath, 'utf-8')) as AppConfig
+    } catch {
+        return DEFAULT_CONFIG
+    }
+}
+
+/**
+ * Сохраняет конфигурацию в файл.
+ *
+ * @param {AppConfig} config - Объект конфигурации для сохранения.
+ */
+export function saveConfig(config: AppConfig): void {
+    fs.writeFileSync(configPath, JSON.stringify(config, null, 2))
+}
